+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Fix #3047: NPE when getting version when there is no build date
 * Fix #3024: stopAllRegisteredInformers will not call startWatcher
 * Fix #3067: Added a patch(PatchContext, item) operation to be more explicit about patching and diffing behavior
+* Fix #3097: refresh token with autoconfigure even if authprovider is null
 
 #### Improvements
 * Fix #2788: Support FIPS mode in kubernetes-client with BouncyCastleFipsProvider